Separate functions and commands folders
