+++ Functions/fortune.js
