+++ Functions/generateQR.js
